requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
archspec
asttokens==2.4.1
astunparse==1.6.3
Babel==2.14.0
basicsr==1.3.4.9
bidict==0.23.1
blinker==1.8.2
boltons
Brotli
build==1.0.3
cachetools==5.3.3
camera==1.3.0
certifi
cffi
chardet==5.2.0
charset-normalizer
check-manifest==0.49
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coverage==7.4.1
cryptography==42.0.7
cryptograpy==0.0.0
cycler==0.12.1
Cython==3.0.9
cython-bbox==0.1.5
dataclasses-json==0.6.4
decorator==5.1.1
distro
dnspython==1.16.0
easydict==1.12
easyocr==1.7.1
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
future
gast==0.4.0
gcloud==0.18.3
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
griffe==0.40.1
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httplib2==0.22.0
hydra-core==1.3.2
idna
imageio==2.34.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonpatch
jsonpointer==2.1
jws==0.1.3
keras==2.11.0
kiwisolver==1.4.5
lapx==0.5.7
lazy_loader==0.4
libclang==16.0.6
libmambapy
loguru
lmdb==1.4.1
loguru==0.7.2
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
menuinst
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==0.5.0
mkdocs-material==9.5.9
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.0
mkdocstrings-python==1.8.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.2
onemetric
omegaconf==2.3.0
onemetric==0.1.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging 
paginate==0.5.6
pandas==2.2.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
Pillow==9.4.0
platformdirs
pluggy
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycosat
pycparser
pycryptodome==3.20.0
Pygments==2.17.2
pymdown-extensions==10.7
pymongo==4.7.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
Pyrebase4==4.5.0
PySocks
pytest==8.0.1
pytest-cov==4.1.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-jwt==2.0.1
python-socketio==5.11.2
pytz==2024.1
PyYAML
pyyaml_env_tag==0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
ruamel.yaml
ruamel.yaml.clib
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy
seaborn==0.13.2
shapely==2.0.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tb-nightly==2.17.0a20240423
tensorboard==2.11.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2024.4.18
tomli==2.0.1
torch==2.3.0+cpu
torchaudio==2.3.0+cpu
torchvision==0.18.0+cpu
tqdm 
traitlets==5.14.1
triton==2.2.0
typing-inspect==0.9.0
typing_extensions
tzdata==2024.1
urllib3==1.26.18
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yapf==0.40.2
zipp==3.17.0
zstandard
